% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonet_prediction_stacking.R
\name{nonet_prediction_stacking}
\alias{nonet_prediction_stacking}
\title{Create a list of predictions}
\usage{
nonet_prediction_stacking(...)
}
\arguments{
\item{...}{arguments supplied to the functions}
}
\value{
A list of ensembled predictions. This list can be used to provide as an input argument for nonet_ensemble function
}
\description{
Create a list of predictions
}
\examples{
# Tune models using only the first 40 rows to keep computation fast

models <- machine_learn(pima_diabetes[1:40, ], patient_id,
                        outcome = diabetes, tune = FALSE)

# Make prediction on the next 10 rows. This uses the best-performing model from
# tuning cross validation, and it also prepares the new data in the same way as
# the training data was prepared.

predictions <- predict(models, newdata = pima_diabetes[41:50, ])
predictions
evaluate(predictions)
plot(predictions)
}
