% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonet_ensemble.R
\name{nonet_ensemble}
\alias{nonet_ensemble}
\title{Ensemble Prediction without using training labels}
\usage{
nonet_ensemble(object, best_modelname)
}
\arguments{
\item{object}{prediction_list object, as from `tune_models`}

\item{best_modelname}{Best model name is one which performance better while evaluating using any evaluation matrix like confusion matrix.}
}
\value{
A list of ensembled predictions. You can evaluate the performance of ensembled prediction using the evaulation matrix as Confusion matrix or AUROC.
}
\description{
Ensemble Prediction without using training labels
}
\examples{
# Tune models using only the first 40 rows to keep computation fast

models <- machine_learn(pima_diabetes[1:40, ], patient_id,
                        outcome = diabetes, tune = FALSE)

# Make prediction on the next 10 rows. This uses the best-performing model from
# tuning cross validation, and it also prepares the new data in the same way as
# the training data was prepared.

predictions <- predict(models, newdata = pima_diabetes[41:50, ])
predictions
evaluate(predictions)
plot(predictions)
}
