% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonet_plot.R
\name{nonet_plot}
\alias{nonet_plot}
\title{Plot the predictions or results of nonet_ensemble}
\usage{
nonet_plot(x, y, dataframe, plot_type, nonet_color, nonet_size,
  nonet_alpha, nonet_bins)
}
\arguments{
\item{x}{}

\item{y}{}

\item{plot_type}{}

\item{nonet_color}{}

\item{nonet_size}{}

\item{nonet_alpha}{}

\item{nonet_bins}{}

\item{data}{}
}
\value{
A list of ensembled predictions. This list can be used to provide as an input argument for nonet_ensemble function
}
\description{
Plot the predictions or results of nonet_ensemble
}
\examples{
# Tune models using only the first 40 rows to keep computation fast

models <- machine_learn(pima_diabetes[1:40, ], patient_id,
                        outcome = diabetes, tune = FALSE)

# Make prediction on the next 10 rows. This uses the best-performing model from
# tuning cross validation, and it also prepares the new data in the same way as
# the training data was prepared.

predictions <- predict(models, newdata = pima_diabetes[41:50, ])
predictions
evaluate(predictions)
plot(predictions)
}
